+++ CHANGELOG.md
@@ -17,3 +17,4 @@ Unreleased Changes
 - Directly commit into a new git instance
 - Split the wiki translator out into a separate module
 - Rebuild the translator to use pandoc on a preprocessed html fragment
+- Add a synthetic home page to generated wiki